modify some divers
